requirements.txt
Flask==2.2.5
transformers==4.32.0
torch==2.0.1
google-generativeai==0.8.4  